move DocValuesRangeIterator from document -> search pkg
